+++ doc/whats_new.rst
@@ -47,10 +47,10 @@ Enhancements
   to :func:`xarray.Dataset.xsimlab.run`.
 - More consistent dictionary format for output variables in the xarray
   extension (:issue:`85`).
-- :func:`~xsimlab.foreign` has been updated to get the original description
-  of a foreign variable
 - %-formatting and str.format() code has been converted into formatted string
   literals (f-strings) (:issue:`90`).
+- :func:`~xsimlab.foreign` has been updated to get the original description
+  of a foreign variable (:issue:`91`)
 
 Bug fixes
 ~~~~~~~~~